third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1251_include=css-expected.txt

This is a testharness.js-based test.
[FAIL] CSS <link> (windows-1251) #<id> { background-image:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> #<id> { background-image:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <link> (windows-1251) #<id> { border-image-source:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> #<id> { border-image-source:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <link> (windows-1251) #<id>::before { content:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> #<id>::before { content:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <link> (windows-1251) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <link> (windows-1251) #<id> { display:list-item; list-style-image:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> #<id> { display:list-item; list-style-image:<url> }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <link> (windows-1251) #<id> { cursor:<url>, pointer }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
[FAIL] CSS <style> #<id> { cursor:<url>, pointer }
  assert_equals: expected "%C3%A5" but got "%26%23229%3B"
Harness: the test ran to completion.
